net/rds: RDS connection shutdown stuck after CQ access violation error
The HCA has reported a CQ access error and RDS tries to drop the connection
to reconnect again. But the RDS shutdown is stuck indefinitely in the
rds_ib_conn_path_shutdown() fn looping in wait_event_timeout().

This patch adds the following changes:

1. rds_ib_cq_event_handler() has received the error from HCA, sets
RDS_IB_CQ_ERR and wakes up the wait_queue. It requires memory barrier
before calling waitqueue_active. Adding the set_bit() helper function
takes care of it.

2. When the CQ is in error state, the same CQ cannot be reused.
So just destroy this CQ and RDS reconnect will create a new CQ.

3. Adds a new cq_event handler for the CQ created by fastreg as it uses
rds_ibdev object which is not supported by the current cq_event handler.

4. Converts bit operations on ic->i_flags to use set|clear_bit() API with
a helper function suggested by Haakon.

The functionality handling CQ events has been tested by the
undersigned on an IB (CX-3) system. An error injecting method was
applied; reducing the number of entries in the CQ by executing
ib_resize_cq() on it. This has been done on the data-plane CQs and on
the CQ for the FRWR (forced RDS to use FRWR by means of the
prefer_frwr module parameter).

Orabug: 33811824

Fixes: e9595a7b949f (net/rds: RDS connection does not reconnect after CQ access violation error)
Fixes: 738c21813c9d (net/rds: Release all resources allocated by "rds_ib_setup_qp")

Suggested-by: Håkon Bugge <haakon.bugge@oracle.com>
Tested-by: Aru Kolappan <aru.kolappan@oracle.com>

+++ net/rds/ib.h
@@ -305,7 +305,7 @@ struct rds_ib_connection {
 	u16			i_frag_sz;	/* IB fragment size */
 	u16			i_frag_cache_sz;
 	u8			i_frag_pages;
-	u8			i_flags;
+	unsigned long		i_flags;
 	u16			i_frag_cache_inx;
 	u16			i_hca_sge;
